+++ src/popup/components/Popups/RebindGroupPopup.vue
@@ -4,8 +4,8 @@ import { useGroupStore } from '@/stores/group'
 import { trans } from '@common/modules/trans'
 import { usePopupStore } from '@/stores/popup'
 import { useRouter } from 'vue-router'
-import { validateURL } from '@common/modules/validateURL'
-import { hashURL } from '@/modules/hashURL'
+import { validateURL } from '@/modules/url/validateURL'
+import { hashURL } from '@/modules/url/hashURL'
 import { error } from '@common/modules/error'
 import { showToast } from '@common/modules/showToast'
 import Popup from '@/components/Popups/Popup.vue'
